--- src/pages/user/components/LineOr.scss
@@ -1,32 +1,0 @@
-.line_or {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    position: relative;
-    width: 100%;
-    height: 1rem;
-    padding: 15px 0;
-}
-
-.line_or span {
-    background-color: #f3f3f4;
-    padding: 0 1rem;
-    z-index: 1;
-    position: relative;
-}
-
-.line_or::before,
-.line_or::after {
-    content: '';
-    background-color: #000;
-    height: 1px;
-    flex: 1;
-}
-
-.line_or::before {
-    margin-right: 0.5rem;
-}
-
-.line_or::after {
-    margin-left: 0.5rem;
-}
