--- src/pages/user/components/LineOr.tsx
@@ -1,11 +1,0 @@
-import './LineOr.scss';
-
-const LineOr = () => {
-    return (
-        <span className="line_or">
-            <span>또는</span>
-        </span>
-    );
-};
-
-export default LineOr;
